+++ src/main/java/com/liveeasystreet/ecovalue/dto/QuizDTO.java
@@ -1,6 +1,12 @@
 package com.liveeasystreet.ecovalue.dto;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
 public class QuizDTO {
-
-
+    private String category;
+    private String title;
+    private String detail;
+    private String solve;
 }
